move ContractState -> types dir, use existing WinType enum
