--- packages/frontend/src/blockchain/abi/WinType.ts
@@ -1,6 +1,0 @@
-export enum WinType {
-  LOSS,
-  GOLDEN_TICKET,
-  AUCTION,
-  RAFFLE,
-}
