+++ packages/frontend/src/components/userActious/getWarningText.ts
@@ -1,4 +1,4 @@
-import { ContractState } from '@/blockchain/abi/ContractState'
+import { ContractState } from '@/types/ContractState'
 
 export function getWarningText(state: ContractState) {
   if (state === ContractState.RAFFLE_SETTLED) {
