+++ packages/frontend/src/utils/isAuctionSettled.ts
@@ -1,4 +1,4 @@
-import { ContractState } from '@/blockchain/abi/ContractState'
+import { ContractState } from '@/types/ContractState'
 
 export function isAuctionSettled(state: ContractState) {
   return [ContractState.AUCTION_SETTLED, ContractState.RAFFLE_SETTLED, ContractState.CLAIMING_CLOSED].includes(state)
